* [#16] Moved ArgyllCMS related utility functions to their own module at `DisplayCAL.argyll`.
